% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdautils.R
\name{na.omit.unique}
\alias{na.omit.unique}
\title{Removes NA from a feature}
\usage{
na.omit.unique(feature)
}
\arguments{
\item{feature}{the feature to remove NA values from}
}
\value{
the non-NA unique values of the input feature
}
\description{
Takes a features of a dataset and removes every NA value, returning only the the unique feature values
}
